Add regression test for the video URL after state restoration
Moves FakeFileRepository out of VideoUrlResolverTest so both tests share it.
